plugin/scripts/requirements.txt
requests>=2.28.0
tweepy>=4.14.0
pyyaml>=6.0
